# documentation/swagger_collection.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Authorization service AUTH",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/ms-auth",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/token/create": {
      "post": {
        "tags": [
          "The TokenController API"
        ],
        "summary": "Creates a new pair of access and refresh tokens",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokensRq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "502": {
            "description": "Bad Gateway",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTokensRs"
                }
              }
            }
          }
        }
      }
    },
    "/token/{refreshToken}/refresh": {
      "get": {
        "tags": [
          "The TokenController API"
        ],
        "summary": "Refreshes tokens for user by his refresh token",
        "operationId": "refresh",
        "parameters": [
          {
            "name": "refreshToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "502": {
            "description": "Bad Gateway",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/token/{accessToken}/validate": {
      "get": {
        "tags": [
          "The TokenController API"
        ],
        "summary": "Validating given access token",
        "operationId": "validate",
        "parameters": [
          {
            "name": "accessToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "502": {
            "description": "Bad Gateway",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Unit"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorDetails": {
        "required": [
          "exceptionClass",
          "exceptionProducer",
          "message",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "message": {
            "type": "string"
          },
          "exceptionClass": {
            "type": "string"
          },
          "exceptionProducer": {
            "type": "string"
          }
        }
      },
      "CreateTokensRq": {
        "required": [
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "CreateTokensRs": {
        "required": [
          "accessToken",
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "Unit": {
        "type": "object"
      }
    }
  }
}